+++ packages/office-ui-fabric-react/src/components/ComboBox/VirtualizedComboBox.tsx
@@ -1,13 +1,14 @@
 import * as React from 'react';
 import { autobind, BaseComponent, customizable } from '../../Utilities';
 import { ComboBox } from './ComboBox';
+import { ComboBoxBase } from './ComboBox.base';
 import { IComboBoxProps, IComboBox } from './ComboBox.types';
 import { List } from '../../List';
 import { SelectableOptionMenuItemType, ISelectableOption } from '../../utilities/selectableOption/SelectableOption.types';
 
 export class VirtualizedComboBox extends BaseComponent<IComboBoxProps, {}> implements IComboBox {
   /** The combo box element */
-  private _comboBox: ComboBox;
+  private _comboBox: ComboBoxBase;
   /** The virtualized list element */
   private _list: List;
