+++ packages/office-ui-fabric-react/src/components/ComboBox/index.ts
@@ -1,4 +1,5 @@
 export { SelectableOptionMenuItemType } from '../../utilities/selectableOption/SelectableOption.types';
 export * from './ComboBox';
+export * from './ComboBox.base';
 export * from './ComboBox.types';
 export * from './VirtualizedComboBox';
